Simplify token retrieval
